+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased][Unreleased]
+## [0.4.0][0.4.0] - 2025-06-26
 
 * Fixed an issue where a project reference (blue) could be wrongly identified as package reference (green)
 * Removed the `--no-browser` option, replaced with the `-u|--url` option
@@ -16,5 +16,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Initial release
 
-[Unreleased]: https://github.com/0xced/nugraph/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/0xced/nugraph/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/0xced/nugraph/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/0xced/nugraph/compare/b581197c8849922788f3e79fd88b417a8ca18db6...0.3.0